# docs/swagger.yaml
definitions:
...
  api.EmptyResponse:
    type: object
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.SuccessResponse:
    properties:
      success:
        type: boolean
    type: object
...
  database.Item:
    properties:
      ItemId:
        type: string
      SlotIndex:
        type: integer
      StackCount:
        type: integer
    type: object
  database.Items:
    properties:
      CommonContainerId:
        items:
          $ref: '#/definitions/database.Item'
        type: array
      DropSlotContainerId:
        items:
          $ref: '#/definitions/database.Item'
        type: array
      EssentialContainerId:
        items:
          $ref: '#/definitions/database.Item'
        type: array
      FoodEquipContainerId:
        items:
          $ref: '#/definitions/database.Item'
        type: array
      PlayerEquipArmorContainerId:
        items:
          $ref: '#/definitions/database.Item'
        type: array
      WeaponLoadOutContainerId:
        items:
          $ref: '#/definitions/database.Item'
        type: array
    type: object
  database.OnlinePlayer:
    properties:
      ip:
        type: string
      last_online:
        type: string
      level:
        type: integer
      location_x:
        type: number
      location_y:
        type: number
      nickname:
        type: string
      ping:
        type: number
      player_uid:
        type: string
      steam_id:
        type: string
    type: object
  database.Pal:
    properties:
      defense:
        type: integer
      exp:
        type: integer
      gender:
        type: string
      hp:
        type: integer
      is_boss:
        type: boolean
      is_lucky:
        type: boolean
      is_tower:
        type: boolean
      level:
        type: integer
      max_hp:
        type: integer
      melee:
        type: integer
      nickname:
        type: string
      ranged:
        type: integer
      rank:
        type: integer
      rank_attack:
        type: integer
      rank_craftspeed:
        type: integer
      rank_defence:
        type: integer
      skills:
        items:
          type: string
        type: array
      type:
        type: string
      workspeed:
        type: integer
    type: object
  database.Player:
    properties:
      exp:
        type: integer
      full_stomach:
        type: number
      hp:
        type: integer
      ip:
        type: string
      items:
        $ref: '#/definitions/database.Items'
      last_online:
        type: string
      level:
        type: integer
      location_x:
        type: number
      location_y:
        type: number
      max_hp:
        type: integer
      max_status_point:
        type: integer
      nickname:
        type: string
      pals:
        items:
          $ref: '#/definitions/database.Pal'
        type: array
      ping:
        type: number
      player_uid:
        type: string
      save_last_online:
        type: string
      shield_hp:
        type: integer
      shield_max_hp:
        type: integer
      status_point:
        additionalProperties:
          type: integer
        type: object
      steam_id:
        type: string
    type: object
  database.PlayerW:
    properties:
      name:
        type: string
      player_uid:
        type: string
      steam_id:
        type: string
    type: object
...
  database.TersePlayer:
    properties:
      exp:
        type: integer
      full_stomach:
        type: number
      hp:
        type: integer
      ip:
        type: string
      last_online:
        type: string
      level:
        type: integer
      location_x:
        type: number
      location_y:
        type: number
      max_hp:
        type: integer
      max_status_point:
        type: integer
      nickname:
        type: string
      ping:
        type: number
      player_uid:
        type: string
      save_last_online:
        type: string
      shield_hp:
        type: integer
      shield_max_hp:
        type: integer
      status_point:
        additionalProperties:
          type: integer
        type: object
      steam_id:
        type: string
    type: object
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /api/online_player:
    get:
      consumes:
      - application/json
      description: List Online Players
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.OnlinePlayer'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List Online Players
      tags:
      - Player
  /api/player:
    get:
      consumes:
      - application/json
      description: List Players
      parameters:
      - description: order by field
        enum:
        - last_online
        - level
        in: query
        name: order_by
        type: string
      - description: order by desc
        in: query
        name: desc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.TersePlayer'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List Players
      tags:
      - Player
    put:
      consumes:
      - application/json
      description: Put Players Only For SavSync,PlayerSync
      parameters:
      - description: Players
        in: body
        name: players
        required: true
        schema:
          items:
            $ref: '#/definitions/database.Player'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Put Players
      tags:
      - Player
  /api/player/{player_uid}:
    get:
      consumes:
      - application/json
      description: Get Player
      parameters:
      - description: Player UID
        in: path
        name: player_uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Player'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.EmptyResponse'
      summary: Get Player
      tags:
      - Player
  /api/player/{player_uid}/ban:
    post:
      consumes:
      - application/json
      description: Ban Player
      parameters:
      - description: Player UID
        in: path
        name: player_uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Ban Player
      tags:
      - Player
  /api/player/{player_uid}/kick:
    post:
      consumes:
      - application/json
      description: Kick Player
      parameters:
      - description: Player UID
        in: path
        name: player_uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Kick Player
      tags:
      - Player
  /api/player/{player_uid}/unban:
    post:
      consumes:
      - application/json
      description: Unban Player
      parameters:
      - description: Player UID
        in: path
        name: player_uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Unban Player
      tags:
      - Player
...
  /api/whitelist:
    delete:
      consumes:
      - application/json
      description: Remove White List
      parameters:
      - description: Player UID
        in: path
        name: player_uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Remove White List
      tags:
      - Player
    get:
      consumes:
      - application/json
      description: List White List
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.PlayerW'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List White List
      tags:
      - Player
    post:
      consumes:
      - application/json
      description: Add White List
      parameters:
      - description: Player UID
        in: path
        name: player_uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add White List
      tags:
      - Player
    put:
      consumes:
      - application/json
      description: Put White List
      parameters:
      - description: Players
        in: body
        name: players
        required: true
        schema:
          items:
            $ref: '#/definitions/database.PlayerW'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Put White List
      tags:
      - Player
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"